New repositiries, services and domains for address and accumulative card

+++ src/main/java/com/lab/delivery/domain/Pizza.java
@@ -6,10 +6,10 @@ package com.lab.delivery.domain;
 public class Pizza {
     private Integer id;
     private String name;
-    private Double price;
+    private Integer price;
     private PizzaTipe pizzaTipe;
 
-    public Pizza(Integer id, String name, Double price, PizzaTipe pizzaTipe) {
+    public Pizza(Integer id, String name, Integer price, PizzaTipe pizzaTipe) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -32,11 +32,11 @@ public class Pizza {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
